refactor: producer
